Let SharedMemory and NDArray support `with` blocks
To do this, we subclass Python's SharedMemory class,
rather than simply passing it along verbatim anymore.

And we implement the __exit__ method to call the Appose SharedMemory
subclass's dispose() method, which calls either close() or unlink()
depending on whether the unlink_on_dispose flag is set.

Note that this implementation cannot quite align with the Java one,
because in Java, the AutoCloseable interface always calls close().
As such, the close() method must be overridden and taught to sometimes
call unlink(), and sometimes not, depending on the unlinkOnClose flag.