Handle sys.exit properly in python_worker
Weirdly, when sys.exit is called in a Python thread, it
does not exit the process, but rather only kills the thread.
To detect this situation, we start a dedicated Appose-Janitor
thread for monitoring existing tasks for dead threads, and
issue failure responses for any that did not already issue a
terminal response normally (presumably because they were killed).