Enable dummy worker Task objects with no-args ctor
Allowing this makes the following magic header possible:

    from appose.python_worker import Task
    task = globals().get("task", Task())

placed at the top of an Appose task script, so that the script can be
more easily developed and tested within an IDE "outside of" Appose.

When the task object has been injected, it's a no-op, but if the
script is executing standalone, the task will be a dummy one.